feat(gitlab-mcp): complete modular architecture refactoring
Transform monolithic 2000+ line server.py into clean modular structure:

Modules created:
- pipeline_handlers.py: 7 CI/CD pipeline debugging tools
- project_handlers.py: 10 project/branch/tag management tools  
- basic_handlers.py: 5 core GitLab operations (with #756 fixes)
- user_group_handlers.py: 7 user/group management tools
- issue_handlers.py: 10 comprehensive issue management tools
- merge_request_handlers.py: 11 advanced merge request tools
- utils.py: Shared utilities and enhanced error handling

Benefits:
- Maintainable: Clear separation of concerns
- Extensible: Easy to add new tools to appropriate modules
- Debuggable: Focused modules make issues easier to trace
- Readable: Each module has single responsibility

All 53 tools maintain exact same functionality and schemas.
Fixes #756 unknown flags and JSON parsing errors.

Principle: systems-stewardship
